+++ taktik/core/social_media/instagram/actions/business/workflows/post_url/workflow.py
@@ -22,7 +22,7 @@ class PostUrlBusiness(
     def __init__(self, device, session_manager=None, automation=None):
         super().__init__(device, session_manager, automation, "post-url", init_business_modules=True)
         from ...common.workflow_defaults import POST_URL_DEFAULTS
-        from .....ui.selectors import HASHTAG_SELECTORS
+        from taktik.core.social_media.instagram.ui.selectors.surfaces.hashtag import HASHTAG_SELECTORS
         self.default_config = {**POST_URL_DEFAULTS}
         self._hashtag_sel = HASHTAG_SELECTORS